#41078 Delete upgrade method for migrations
